move default_colors.py from folder "scss" to folder "modules"
