+++ modules/default_colors.py
